Conflicts:
	grails-app/domain/th/co/opendream/cashcard/domain/InterestRate.groovy

+++ grails-app/domain/th/co/opendream/cashcard/domain/InterestRate.groovy
@@ -2,12 +2,12 @@ package th.co.opendream.cashcard.domain
 
 class InterestRate {
     Date startDate
-    Float rate
+    BigDecimal rate
     Date dateCreated
     Date lastUpdated
 
     static constraints = {
         startDate unique: true, nullable: false, blank: false
-        rate min: 0F, nullable: false, blank: false
+        rate min: 0.00, nullable: false, blank: false
     }
 }
